# docs/swagger/swagger.yaml
basePath: /
definitions:
  helper.JsonObject:
    properties:
      code:
        description: '* 状态码'
        type: string
      content:
        description: '* 内容体'
        type: object
      message:
        description: '* 消息'
        type: string
    type: object
  helper.PageBean:
    properties:
      page:
        description: '* 当前页'
        type: integer
      pageSize:
        description: '* 每页显示的最大行数'
        type: integer
      rows:
        description: '* 每行的数据'
        type: object
      total:
        description: '* 总记录数'
        type: integer
    type: object
  model.Role:
    properties:
      createdAt:
        description: '* 创建时间'
        type: string
      deletedAt:
        description: '* 删除时间'
        type: string
      description:
        description: '* 角色描述'
        type: string
      id:
        description: '* 主键id'
        type: string
      roleKey:
        description: '* 角色类别标识'
        type: string
      roleName:
        description: '* 角色名称'
        type: string
      updatedAt:
        description: '* 更新时间'
        type: string
    required:
    - roleName
    - roleKey
    type: object
  model.User:
    properties:
      createdAt:
        description: '* 创建时间'
        type: string
      deletedAt:
        description: '* 删除时间'
        type: string
      email:
        description: '* 邮件'
        type: string
      flag:
        description: '* 标志 1 表示这个账号是由管理方为商户添加的账号'
        type: integer
      id:
        description: '* 主键'
        type: string
      loginTime:
        description: '* 最后一次登陆时间'
        type: string
      logonCount:
        description: '* 登陆次数'
        type: integer
      merchantName:
        description: '* 商户名称'
        type: string
      merchantNo:
        description: '* 商户号'
        type: string
      phone:
        description: '* 电话'
        type: string
      role:
        $ref: '#/definitions/model.Role'
        description: '* 用户对应的角色'
        type: object
      roleId:
        description: '* 外键'
        type: string
      status:
        description: '* 状态  0 正常'
        type: integer
      updatedAt:
        description: '* 更新时间'
        type: string
      userName:
        description: '* 姓名'
        type: string
    required:
    - userName
    - phone
    type: object
host: localhost:8080
info:
  contact:
    email: yongtao.yin@bsit.cn
    name: YinYongTao
    url: https://github.com/YinYongTao/gin-first
  description: 这是一个 gin 框架应用 swagger 的示例
  license: {}
  title: gin-first API 文档
  version: "2.0"
paths:
  /api/delete_user:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户记录id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.JsonObject'
            type: object
      summary: 删除用户信息
      tags:
      - UserController
  /api/get_all_users:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.JsonObject'
            type: object
      summary: 获取所有用户信息
      tags:
      - UserController
  /api/get_user:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户记录id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
            type: object
      summary: 用户信息回显
      tags:
      - UserController
  /api/get_user_page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      - description: 每页显示最大行
        in: query
        name: page_size
        required: true
        type: string
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 电话
        in: query
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.PageBean'
            type: object
      summary: 用户信息分页查询
      tags:
      - UserController
  /api/save_role:
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色id,新增时id为空
        in: query
        name: id
        type: string
      - description: 角色名称
        in: query
        name: role_name
        required: true
        type: string
      - description: 角色类别标识
        in: query
        name: role_key
        required: true
        type: string
      - description: 角色描述信息
        in: query
        name: description
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.JsonObject'
            type: object
      summary: 添加修改角色
      tags:
      - RoleController
  /api/save_user:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户记录id,新增时id为空
        in: query
        name: id
        type: string
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      - description: 电话号码
        in: query
        name: phone
        required: true
        type: string
      - description: 邮件
        in: query
        name: email
        required: true
        type: string
      - description: 商户号
        in: query
        name: merchant_no
        required: true
        type: string
      - description: 角色id
        in: query
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.JsonObject'
            type: object
      summary: 添加、修改用户信息
      tags:
      - UserController
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.JsonObject'
            type: object
      summary: 用户登陆接口
      tags:
      - LoginController
swagger: "2.0"